requirements.txt
beautifulsoup4==4.6.3
scrapy-fake-useragent==1.1.0